# Update the query that fetches roles to be faster with many roles