--- a/web/pgadmin/browser/server_groups/servers/roles/templates/roles/sql/default/nodes.sql
+++ b/web/pgadmin/browser/server_groups/servers/roles/templates/roles/sql/default/nodes.sql
@@ -1,8 +1,8 @@
 SELECT
-	r.oid, r.rolname, r.rolcanlogin, r.rolsuper,
-	pg_catalog.shobj_description(r.oid, 'pg_authid') AS description
-FROM
-	pg_catalog.pg_roles r
+    r.oid, r.rolname, r.rolcanlogin, r.rolsuper, d.description
+FROM pg_catalog.pg_roles r
+    LEFT JOIN pg_catalog.pg_shdescription d
+    ON d.objoid = r.oid AND d.classoid = 'pg_catalog.pg_authid'::regclass
 {% if rid %}
 WHERE r.oid = {{ rid|qtLiteral(conn) }}::oid
 {% endif %}